[GT-133] chanage group number of Matcher
http://jira.iitp.cubrid.org/browse/GT-133

After changing the URL_PATTERN string from "jdbc:cubrid(-oracle|-mysql)?:...." to "jdbc:turbograph:......",
the second string was changed to a fixed string not a pattern string.
Because it is a string excluded from the pattern group, the group number for finding each pattern was changed.
ex) change from "host = matcher.group(2)" to "host = matcher.group(1)"